+++ src/main/java/edu/pet/votesystem/model/Restaurant.java
@@ -12,7 +12,7 @@ public class Restaurant {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "rest_id", nullable = false, unique = true)
-    private Integer restaurantId;
+    private Long restaurantId;
     @NotEmpty(message = "restaurant name should not be empty")
     @Size(min = 2, max = 100, message = "restaurant name should be between 2 and 100 characters")
     @Column(name = "rest_name", nullable = false)
@@ -24,11 +24,11 @@ public class Restaurant {
             mappedBy = "restaurant")
     private List<Vote> votes;
 
-    public Integer getRestaurantId() {
+    public Long getRestaurantId() {
         return restaurantId;
     }
 
-    public void setRestaurantId(Integer restaurantId) {
+    public void setRestaurantId(Long restaurantId) {
         this.restaurantId = restaurantId;
     }
